requirements.txt
torch~=1.12.1
wikipedia~=1.4.0
ipython~=8.5.0
transformers~=4.23.1
GoogleNews~=1.6.5
pyvis~=0.3.0
newspaper3k~=0.2.8
coreferee~=1.3.0
spacy~=3.4.2